Make an ember-cli addon.
