+++ .npmignore
@@ -1,0 +1,15 @@
+bower_components/
+tests/
+tmp/
+dist/
+
+.bowerrc
+.editorconfig
+.ember-cli
+.travis.yml
+.npmignore
+**/.gitkeep
+bower.json
+ember-cli-build.js
+Brocfile.js
+testem.json
