+++ .watchmanconfig
@@ -1,0 +1,3 @@
+{
+  "ignore_dirs": ["tmp", "dist"]
+}
