--- Assetfile
@@ -1,39 +1,0 @@
-instance_eval File.read(EmberDev.support_path.join('Assetfile'))
-
-distros = {
-  :full    => %w(ember-resolver)
-}
-
-output "dist"
-
-distros.each do |name, modules|
-  name = "ember-resolver"
-
-  input "dist/modules" do
-    module_paths = modules.map{|m| "#{m}.js" }
-    match "{#{module_paths.join(',')}}" do
-      concat(module_paths){ ["#{name}.js", "#{name}.prod.js"] }
-    end
-
-    match "#{name}.js" do
-      filter VersionInfo
-      filter EmberLicenseFilter
-    end
-
-    # Strip dev code
-    match "#{name}.prod.js" do
-      filter VersionInfo
-      filter EmberLicenseFilter
-      filter(EmberStripDebugMessagesFilter) { ["#{name}.prod.js", "min/#{name}.js"] }
-    end
-
-    # Minify
-    match "min/#{name}.js" do
-      uglify{ "#{name}.min.js" }
-      filter VersionInfo
-      filter EmberLicenseFilter
-    end
-  end
-end
-
-# vim: filetype=ruby
