--- Gemfile
@@ -1,6 +1,0 @@
-source 'https://rubygems.org'
-
-gem 'rake-pipeline', git: 'https://github.com/livingsocial/rake-pipeline.git'
-gem 'ember-dev',     git: 'https://github.com/emberjs/ember-dev.git', branch: 'ruby'
-
-gem 'ember-source', '~> 1.5.0'
