+++ README.md
@@ -1,19 +1,25 @@
-# Ember Resolver [![Build Status](https://travis-ci.org/ember-cli/ember-resolver.svg?branch=master)](https://travis-ci.org/ember-cli/ember-resolver)
+# Ember-resolver
 
-This project is tracking a new resolver based on ES6 semantics that has been extracted from (and used by) the following projects:
+This README outlines the details of collaborating on this Ember addon.
 
-* [ember-cli](https://github.com/ember-cli/ember-cli)
-* [ember-app-kit](https://github.com/stefanpenner/ember-app-kit)
-* [ember-appkit-rails](https://github.com/dockyard/ember-appkit-rails)
+## Installation
 
-## Goal
+* `git clone` this repository
+* `npm install`
+* `bower install`
 
-The goal of this project is to come up with a new default resolver for the next major version of `Ember` (2.0).
+## Running
 
+* `ember server`
+* Visit your app at http://localhost:4200.
 
-## build
+## Running Tests
 
-```
-bundle
-bundle exec rake
-```
+* `ember test`
+* `ember test --server`
+
+## Building
+
+* `ember build`
+
+For more information on using ember-cli, visit [http://www.ember-cli.com/](http://www.ember-cli.com/).
