+++ addon/resolver.js
@@ -1,0 +1,370 @@
+/*globals require */
+
+import Ember from 'ember';
+import entries from './utils/module-entries';
+import classFactory from './utils/class-factory';
+import makeDictionary from './utils/make-dictionary';
+
+/*
+ * This module defines a subclass of Ember.DefaultResolver that adds two
+ * important features:
+ *
+ *  1) The resolver makes the container aware of es6 modules via the AMD
+ *     output. The loader's _moduleEntries is consulted so that classes can be
+ *     resolved directly via the module loader, without needing a manual
+ *     `import`.
+ *  2) is able to provide injections to classes that implement `extend`
+ *     (as is typical with Ember).
+ */
+
+
+let {
+  classify
+} = Ember.String;
+let {
+  get,
+  DefaultResolver
+} = Ember;
+
+function parseName(fullName) {
+  /*jshint validthis:true */
+
+  if (fullName.parsedName === true) { return fullName; }
+
+  var prefix, type, name;
+  var fullNameParts = fullName.split('@');
+
+  // HTMLBars uses helper:@content-helper which collides
+  // with ember-cli namespace detection.
+  // This will be removed in a future release of HTMLBars.
+  if (fullName !== 'helper:@content-helper' &&
+      fullNameParts.length === 2) {
+    var prefixParts = fullNameParts[0].split(':');
+
+    if (prefixParts.length === 2) {
+      prefix = prefixParts[1];
+      type = prefixParts[0];
+      name = fullNameParts[1];
+    } else {
+      var nameParts = fullNameParts[1].split(':');
+
+      prefix = fullNameParts[0];
+      type = nameParts[0];
+      name = nameParts[1];
+    }
+  } else {
+    fullNameParts = fullName.split(':');
+    type = fullNameParts[0];
+    name = fullNameParts[1];
+  }
+
+  var fullNameWithoutType = name;
+  var namespace = get(this, 'namespace');
+  var root = namespace;
+
+  return {
+    parsedName: true,
+    fullName: fullName,
+    prefix: prefix || this.prefix({type: type}),
+    type: type,
+    fullNameWithoutType: fullNameWithoutType,
+    name: name,
+    root: root,
+    resolveMethodName: "resolve" + classify(type)
+  };
+}
+
+function resolveOther(parsedName) {
+  /*jshint validthis:true */
+
+  Ember.assert('`modulePrefix` must be defined', this.namespace.modulePrefix);
+
+  var normalizedModuleName = this.findModuleName(parsedName);
+
+  if (normalizedModuleName) {
+    var defaultExport = this._extractDefaultExport(normalizedModuleName, parsedName);
+
+    if (defaultExport === undefined) {
+      throw new Error(" Expected to find: '" + parsedName.fullName + "' within '" + normalizedModuleName + "' but got 'undefined'. Did you forget to `export default` within '" + normalizedModuleName + "'?");
+    }
+
+    if (this.shouldWrapInClassFactory(defaultExport, parsedName)) {
+      defaultExport = classFactory(defaultExport);
+    }
+
+    return defaultExport;
+  } else {
+    return this._super(parsedName);
+  }
+}
+
+// Ember.DefaultResolver docs:
+//   https://github.com/emberjs/ember.js/blob/master/packages/ember-application/lib/system/resolver.js
+var Resolver = DefaultResolver.extend({
+  resolveOther,
+  parseName,
+  resolveTemplate: resolveOther,
+  pluralizedTypes: null,
+
+  makeToString: function(factory, fullName) {
+    return '' + this.namespace.modulePrefix + '@' + fullName + ':';
+  },
+
+  shouldWrapInClassFactory: function(/* module, parsedName */){
+    return false;
+  },
+
+  init: function() {
+    this._super();
+    this.moduleBasedResolver = true;
+    this._normalizeCache = makeDictionary();
+
+    this.pluralizedTypes = this.pluralizedTypes || makeDictionary();
+
+    if (!this.pluralizedTypes.config) {
+      this.pluralizedTypes.config = 'config';
+    }
+
+    this._deprecatedPodModulePrefix = false;
+  },
+
+  normalize: function(fullName) {
+    return this._normalizeCache[fullName] || (this._normalizeCache[fullName] = this._normalize(fullName));
+  },
+
+  _normalize: function(fullName) {
+    // replace `.` with `/` in order to make nested controllers work in the following cases
+    // 1. `needs: ['posts/post']`
+    // 2. `{{render "posts/post"}}`
+    // 3. `this.render('posts/post')` from Route
+    var split = fullName.split(':');
+    if (split.length > 1) {
+      return split[0] + ':' + Ember.String.dasherize(split[1].replace(/\./g, '/'));
+    } else {
+      return fullName;
+    }
+  },
+
+  pluralize: function(type) {
+    return this.pluralizedTypes[type] || (this.pluralizedTypes[type] = type + 's');
+  },
+
+  podBasedLookupWithPrefix: function(podPrefix, parsedName) {
+    var fullNameWithoutType = parsedName.fullNameWithoutType;
+
+    if (parsedName.type === 'template') {
+      fullNameWithoutType = fullNameWithoutType.replace(/^components\//, '');
+    }
+
+    return podPrefix + '/' + fullNameWithoutType + '/' + parsedName.type;
+  },
+
+  podBasedModuleName: function(parsedName) {
+    var podPrefix = this.namespace.podModulePrefix || this.namespace.modulePrefix;
+
+    return this.podBasedLookupWithPrefix(podPrefix, parsedName);
+  },
+
+  podBasedComponentsInSubdir: function(parsedName) {
+    var podPrefix = this.namespace.podModulePrefix || this.namespace.modulePrefix;
+    podPrefix = podPrefix + '/components';
+
+    if (parsedName.type === 'component' || parsedName.fullNameWithoutType.match(/^components/)) {
+      return this.podBasedLookupWithPrefix(podPrefix, parsedName);
+    }
+  },
+
+  mainModuleName: function(parsedName) {
+    // if router:main or adapter:main look for a module with just the type first
+    var tmpModuleName = parsedName.prefix + '/' + parsedName.type;
+
+    if (parsedName.fullNameWithoutType === 'main') {
+      return tmpModuleName;
+    }
+  },
+
+  defaultModuleName: function(parsedName) {
+    return parsedName.prefix + '/' +  this.pluralize(parsedName.type) + '/' + parsedName.fullNameWithoutType;
+  },
+
+  prefix: function(parsedName) {
+    var tmpPrefix = this.namespace.modulePrefix;
+
+    if (this.namespace[parsedName.type + 'Prefix']) {
+      tmpPrefix = this.namespace[parsedName.type + 'Prefix'];
+    }
+
+    return tmpPrefix;
+  },
+
+  /**
+
+   A listing of functions to test for moduleName's based on the provided
+   `parsedName`. This allows easy customization of additional module based
+   lookup patterns.
+
+   @property moduleNameLookupPatterns
+   @returns {Ember.Array}
+   */
+  moduleNameLookupPatterns: Ember.computed(function(){
+    return Ember.A([
+      this.podBasedModuleName,
+      this.podBasedComponentsInSubdir,
+      this.mainModuleName,
+      this.defaultModuleName
+    ]);
+  }),
+
+  findModuleName: function(parsedName, loggingDisabled){
+    var self = this;
+    var moduleName;
+
+    this.get('moduleNameLookupPatterns').find(function(item) {
+      var moduleEntries = entries;
+      var tmpModuleName = item.call(self, parsedName);
+
+      // allow treat all dashed and all underscored as the same thing
+      // supports components with dashes and other stuff with underscores.
+      if (tmpModuleName) {
+        tmpModuleName = self.chooseModuleName(moduleEntries, tmpModuleName);
+      }
+
+      if (tmpModuleName && moduleEntries[tmpModuleName]) {
+        if (!loggingDisabled) {
+          self._logLookup(true, parsedName, tmpModuleName);
+        }
+
+        moduleName = tmpModuleName;
+      }
+
+      if (!loggingDisabled) {
+        self._logLookup(moduleName, parsedName, tmpModuleName);
+      }
+
+      return moduleName;
+    });
+
+    return moduleName;
+  },
+
+  chooseModuleName: function(moduleEntries, moduleName) {
+    var underscoredModuleName = Ember.String.underscore(moduleName);
+
+    if (moduleName !== underscoredModuleName && moduleEntries[moduleName] && moduleEntries[underscoredModuleName]) {
+      throw new TypeError("Ambiguous module names: `" + moduleName + "` and `" + underscoredModuleName + "`");
+    }
+
+    if (moduleEntries[moduleName]) {
+      return moduleName;
+    } else if (moduleEntries[underscoredModuleName]) {
+      return underscoredModuleName;
+    } else {
+      // workaround for dasherized partials:
+      // something/something/-something => something/something/_something
+      var partializedModuleName = moduleName.replace(/\/-([^\/]*)$/, '/_$1');
+
+      if (moduleEntries[partializedModuleName]) {
+        Ember.deprecate('Modules should not contain underscores. ' +
+                        'Attempted to lookup "'+moduleName+'" which ' +
+                        'was not found. Please rename "'+partializedModuleName+'" '+
+                        'to "'+moduleName+'" instead.', false);
+
+        return partializedModuleName;
+      } else {
+        return moduleName;
+      }
+    }
+  },
+
+  // used by Ember.DefaultResolver.prototype._logLookup
+  lookupDescription: function(fullName) {
+    var parsedName = this.parseName(fullName);
+
+    var moduleName = this.findModuleName(parsedName, true);
+
+    return moduleName;
+  },
+
+  // only needed until 1.6.0-beta.2 can be required
+  _logLookup: function(found, parsedName, description) {
+    if (!Ember.ENV.LOG_MODULE_RESOLVER && !parsedName.root.LOG_RESOLVER) {
+      return;
+    }
+
+    var symbol, padding;
+
+    if (found) { symbol = '[✓]'; }
+    else       { symbol = '[ ]'; }
+
+    if (parsedName.fullName.length > 60) {
+      padding = '.';
+    } else {
+      padding = new Array(60 - parsedName.fullName.length).join('.');
+    }
+
+    if (!description) {
+      description = this.lookupDescription(parsedName);
+    }
+
+    Ember.Logger.info(symbol, parsedName.fullName, padding, description);
+  },
+
+  knownForType: function(type) {
+    var moduleEntries = entries;
+    var moduleKeys = (Object.keys || Ember.keys)(moduleEntries);
+
+    var items = makeDictionary();
+    for (var index = 0, length = moduleKeys.length; index < length; index++) {
+      var moduleName = moduleKeys[index];
+      var fullname = this.translateToContainerFullname(type, moduleName);
+
+      if (fullname) {
+        items[fullname] = true;
+      }
+    }
+
+    return items;
+  },
+
+  translateToContainerFullname: function(type, moduleName) {
+    var prefix = this.prefix({ type: type });
+
+    // Note: using string manipulation here rather than regexes for better performance.
+    // pod modules
+    // '^' + prefix + '/(.+)/' + type + '$'
+    var podPrefix = prefix + '/';
+    var podSuffix = '/' + type;
+    var start = moduleName.indexOf(podPrefix);
+    var end = moduleName.indexOf(podSuffix);
+
+    if (start === 0 && end === (moduleName.length - podSuffix.length) &&
+        moduleName.length > (podPrefix.length + podSuffix.length)) {
+      return type + ':' + moduleName.slice(start + podPrefix.length, end);
+    }
+
+    // non-pod modules
+    // '^' + prefix + '/' + pluralizedType + '/(.+)$'
+    var pluralizedType = this.pluralize(type);
+    var nonPodPrefix = prefix + '/' + pluralizedType + '/';
+
+    if (moduleName.indexOf(nonPodPrefix) === 0 && moduleName.length > nonPodPrefix.length) {
+      return type + ':' + moduleName.slice(nonPodPrefix.length);
+    }
+  },
+
+  _extractDefaultExport: function(normalizedModuleName) {
+    var module = require(normalizedModuleName, null, null, true /* force sync */);
+
+    if (module && module['default']) {
+      module = module['default'];
+    }
+
+    return module;
+  }
+});
+
+Resolver.reopenClass({
+  moduleBasedResolver: true
+});
+
+export default Resolver;
