+++ addon/utils/module-entries.js
@@ -1,0 +1,7 @@
+/*globals requirejs */
+
+if (typeof requirejs.entries === 'undefined') {
+  requirejs.entries = requirejs._eak_seen;
+}
+
+export default requirejs.entries;
