+++ bower.json
@@ -1,29 +1,19 @@
 {
   "name": "ember-resolver",
-  "version": "0.1.21",
-  "license": "MIT",
-  "main": ["dist/ember-resolver.js", "dist/ember-resolver.prod.js", "dist/ember-resolver.min.js"],
-  "ignore": [
-    ".jshintrc",
-    "**/*.txt",
-    ".gitignore",
-    ".jshintrc",
-    ".travis.yml",
-    "Assetfile",
-    "Gemfile",
-    "Gemfile.lock",
-    "LICENSE",
-    "README.md",
-    "Rakefile",
-    "VERSION",
-    "bower.json",
-    "config.ru",
-    "ember-dev.yml",
-    "generators",
-    "packages",
-    "tests"
-  ],
   "dependencies": {
-    "ember": " > 1.5.0-beta.3"
+    "ember": "2.0.0",
+    "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.4",
+    "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
+    "ember-data": "2.0.0",
+    "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.5",
+    "ember-qunit": "0.4.10",
+    "ember-qunit-notifications": "0.0.7",
+    "ember-resolver": "~0.1.20",
+    "jquery": "^1.11.3",
+    "loader.js": "ember-cli/loader.js#3.2.1",
+    "qunit": "~1.18.0"
+  },
+  "resolutions": {
+    "ember": "2.0.0"
   }
 }
