--- ember-dev.yml
@@ -1,12 +1,0 @@
----
-name: Ember Resolver
-assetfile: 'Assetfile'
-testing_suites:
-  default:
-    - "package=all"
-  all:
-    - "package=all&emberchannel=release"
-    - "package=all&emberchannel=beta"
-    - "package=all&emberchannel=canary"
-testing_packages:
-  - ember-resolver
