--- generators/license.js
@@ -1,6 +1,0 @@
-// ==========================================================================
-// Project:   Ember - JavaScript Application Framework
-// Copyright: Copyright 2013 Stefan Penner and Ember App Kit Contributors
-// License:   Licensed under MIT license
-//            See https://raw.github.com/ember-cli/ember-resolver/master/LICENSE
-// ==========================================================================
