--- packages/ember-resolver/lib/main.js
@@ -1,3 +1,0 @@
-require('ember-resolver/core');
-require('ember-resolver/container-debug-adapter');
-require('ember-resolver/initializers');
