--- packages/ember-resolver/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "ember-resolver",
-  "summary": "Ember Resolver Reboot",
-  "description": "Ember Resolver Reboot",
-  "homepage": "https://github.com/ember-cli/ember-resolver",
-  "author": "Stefan Penner",
-  "version": "0.0.1"
-}
