--- tests/ember_configuration.js
@@ -1,95 +1,0 @@
-/*globals ENV QUnit EmberDev */
-
-// START *SLIGHTLY MODIFIED* EAK LOADER
-var define, requireModule, require, requirejs;
-
-(function() {
-  var registry = {}, seen = {};
-
-  define = function(name, deps, callback) {
-    registry[name] = { deps: deps, callback: callback };
-  };
-
-  requirejs = require = requireModule = function(name) {
-    if (seen.hasOwnProperty(name)) { return seen[name]; }
-    seen[name] = {};
-
-    if (!registry[name]) {
-      throw new Error("Could not find module " + name);
-    }
-
-    var mod = registry[name],
-        deps = mod.deps,
-        callback = mod.callback,
-        reified = [],
-        exports;
-
-    for (var i=0, l=deps.length; i<l; i++) {
-      if (deps[i] === 'exports') {
-        reified.push(exports = {});
-      } else {
-        reified.push(requireModule(resolve(deps[i])));
-      }
-    }
-
-    var value = callback.apply(this, reified);
-    return seen[name] = exports || value;
-
-    function resolve(child) {
-      if (child.charAt(0) !== '.') { return child; }
-      var parts = child.split("/");
-      var parentBase = name.split("/").slice(0, -1);
-
-      for (var i=0, l=parts.length; i<l; i++) {
-        var part = parts[i];
-
-        if (part === '..') { parentBase.pop(); }
-        else if (part === '.') { continue; }
-        else { parentBase.push(part); }
-      }
-
-      return parentBase.join("/");
-    }
-  };
-
-  requirejs.entries = registry;
-  requirejs.clear = function(){
-    requirejs.entries = registry = {};
-    seen = {};
-  };
-})();
-// END EAK LOADER
-
-(function() {
-  window.Ember = {
-    testing: true
-  };
-  window.ENV = window.ENV || {};
-
-  // Test for "hooks in ENV.EMBER_LOAD_HOOKS['hookName'] get executed"
-  ENV.EMBER_LOAD_HOOKS = ENV.EMBER_LOAD_HOOKS || {};
-  ENV.EMBER_LOAD_HOOKS.__before_ember_test_hook__ = ENV.EMBER_LOAD_HOOKS.__before_ember_test_hook__ || [];
-  ENV.__test_hook_count__ = 0;
-  ENV.EMBER_LOAD_HOOKS.__before_ember_test_hook__.push(function(object) {
-    ENV.__test_hook_count__ += object;
-  });
-
-  // Handle extending prototypes
-  QUnit.config.urlConfig.push('extendprototypes');
-
-  // var extendPrototypes = QUnit.urlParams.extendprototypes;
-  // ENV['EXTEND_PROTOTYPES'] = !!extendPrototypes;
-
-  // Don't worry about jQuery version
-  ENV['FORCE_JQUERY'] = true;
-
-  if (EmberDev.jsHint) {
-    // jsHint makes its own Object.create stub, we don't want to use this
-    ENV['STUB_OBJECT_CREATE'] = !Object.create;
-  }
-
-  EmberDev.distros = {
-    spade:   'ember-resolver-spade.js',
-    build:   'ember-resolver.js'
-  };
-})();
