+++ tests/helpers/resolver.js
@@ -1,0 +1,11 @@
+import Resolver from 'ember-resolver';
+import config from '../../config/environment';
+
+var resolver = Resolver.create();
+
+resolver.namespace = {
+  modulePrefix: config.modulePrefix,
+  podModulePrefix: config.podModulePrefix
+};
+
+export default resolver;
